+++ js/constants.js
@@ -1,0 +1,16 @@
+const MAX_USER_PUBLICATION = 25;
+const MAX_NUMBER_AVATAR = 6;
+const MIN_NUMBER_LIKES = 15;
+const MAX_NUMBER_LIKES = 200;
+const MAX_NUMBER_COMMENTS = 5;
+const MIN_ELEMENTS = 1;
+
+
+export {
+  MAX_USER_PUBLICATION,
+  MAX_NUMBER_AVATAR,
+  MIN_NUMBER_LIKES,
+  MAX_NUMBER_LIKES,
+  MAX_NUMBER_COMMENTS,
+  MIN_ELEMENTS
+};
