+++ js/mocks.js
@@ -12,11 +12,36 @@ import {
   MIN_NUMBER_LIKES,
   MAX_NUMBER_LIKES,
   MAX_NUMBER_COMMENTS,
-  DESCRIPTION_PHOTOS,
-  COMMENTATORS_MESSAGES,
-  COMMENTATORS_NAMES,
   MIN_ELEMENTS
-} from './variables.js';
+} from './constants.js';
+
+
+const DESCRIPTION_PHOTOS = [
+  'Каталась на лошадке. Высоченный, седло пришлось в прыжке на него закидывать',
+  'Каждый день влюбляюсь в этот город всë больше',
+  'Не знаю, почему мне так нравятся эти пушисто-шуршащие штучки',
+  'Какая всë-таки этом году сказочная зима',
+  'Наверное, можно бесконечно возвращаться в это место и каждый раз вааау',
+  'Вот такой мой маленький уютный рабочий уголок',
+];
+
+const COMMENTATORS_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const COMMENTATORS_NAMES = [
+  'Евсей',
+  'Аннушка',
+  'Святозар',
+  'Элина',
+  'Кавалькада',
+  'Люсетта',
+];
 
 //Объект комментариев
 const getComments = (id) => {
@@ -41,5 +66,6 @@ const getPublishedPhoto = (id) => {
 
 const userGallery = getRandomNumbers(MIN_ELEMENTS, MAX_USER_PUBLICATION).map((id) => getPublishedPhoto(id));
 userGallery;
+
 
 export { getComments };
